refactor getDisplayResults
